requirements.txt
python-dotenv~=1.0.1
pymongo~=4.9.2
bcrypt~=4.2.1
pydantic~=2.10.4
python-jose~=3.3.0
fastapi~=0.115.6
TikTokLive~=6.3.0.post2